# krcg_api/templates/openapi.yaml
openapi: 3.0.3
info:
  title: KRCG API
  version: "{{version}}"
  description: |
    **V:tES** cards and decks based on VEKN resources, including the [TWDA](http://www.vekn.fr/decks/twd.htm).
  contact:
    name: "KRCG API"
    url: https://github.com/lionel-panhaleux/krcg-api
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
...
components:
  schemas:
...
    deck:
      type: object
      required: ["id", "date", "crypt", "library"]
      properties:
        id: { type: string }
        name: { type: string }
        event: { type: string }
        place: { type: string }
        date: { type: string, format: date }
        player: { type: string }
        author: { type: string }
        players_count: { type: integer }
        tournament_format: { type: string }
        score: { type: string }
        comments: { type: string }
        crypt:
          type: object
          required: ["count", "cards"]
          properties:
            count: { type: integer }
            cards:
              type: array
              minLength: 1
              items: { $ref: "#/components/schemas/card_count" }
        library:
          type: object
          required: ["count", "cards"]
          properties:
            count: { type: integer }
            cards:
              type: array
              minLength: 1
              maxLength: 20
              items:
                type: object
                required: ["type", "count", "cards"]
                properties:
                  type: { type: string }
                  count: { type: integer }
                  cards:
                    type: array
                    minLength: 1
                    maxLength: 90
                    items: { $ref: "#/components/schemas/card_count" }
...
    deck_list:
      type: object
      required: ["count", "deck_list"]
      properties:
        count: { type: integer }
        decks:
          type: array
          minLength: 0
          maxLength: 10000
          items: { $ref: "#/components/schemas/deck_summary" }
...
paths:
...
  /card/{name_or_id}:
    get:
      tags: [Card]
      summary: Get card information
      description: |
        Return the matching card information. Also works with **translated** card names.
      parameters:
        - name: name_or_id
          in: path
          required: true
          schema:
            type: string
            description: "Card name or ID"
      responses:
        "200":
          description: Success, returns the card as JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/card" }
              examples: { card: { $ref: "#/components/examples/card" } }
        "404":
          description: Not found
  /complete/{partial_name}:
    get:
      tags: [Card]
      summary: Complete a card name
      description: |
        Return all candidates matching the partial name.
        If Accept-Language is set, will also return available names in that language,
        in addition to the english completions.
      parameters:
        - name: partial_name
          in: path
          required: true
          schema:
            type: string
            description: "Partial card name"
        - name: Accept-Language
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success, returns a JSON array of card names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Card name
              example: ["The unnamed", "Unnatural Disaster"]
        "404":
          description: Not found
  /candidates:
    post:
      tags: [Card]
      summary: Given a list of cards, returns likely additional candidates for a deck
      description: |
        Only 10 candidates are returned. If there less than 4 decks in the TWDA
        tha ressemble the cards list provided, 404 is returned.
        If no card is provided, it will simply output the list of most played cards.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  description: response format (return name or full card).
                  enum: ["name", "full"]
                players_count:
                  type: integer
                  description: |
                    If set, only consider tournament winning decks from tournaments
                    with at least that many players participating
                  default: 0
                date_from:
                  type: string
                  format: date
                  description: |
                    If set, only consider tournament winning decks from that date on
                date_to:
                  type: string
                  format: date
                  description: |
                    If set, only consider tournament winning decks up to that date
                cards:
                  type: array
                  items:
                    oneOf:
                      - type: string
                        description: Card name
                      - type: integer
                        description: Card ID
            example:
              cards: [Cybele, Nana Buruku]
              date_from: "2015"
              date_to: "2020"
      responses:
        "200":
          description: Success, returns a list of up to 10 candidates
          content:
            application/json:
              schema:
                type: array
                description: |
                  Card names, or full card JSON objects
                  if the "full" mode has been requested
                items:
                  type: object
                  properties:
                    card:
                      description: |
                        Card name or JSON, depending on the "mode"
                        provided in the request.
                      oneOf:
                        - type: string
                          description: Card name
                        - { $ref: "#/components/schemas/card" }
                    score:
                      type: number
                      description: |
                        Percentage of presence in similar tournament winning decks: 0.4
                        means the card is listed in 40% of similar decks in the TWDA.
                    average:
                      type: integer
                      description: |
                        Average number of copies used in similar decks using the card.
                    deviation:
                      type: number
                      description: |
                        Standard deviation for the number of copies.
                        Note that average has been rounded (so presents a 0.5 error).
                        A "typical" range for copies can be computed using:
                        floor(max(1, average - deviation)), ceil(average + deviation)
              example:
                - {
                    "average": 14,
                    "card": "Ashur Tablets",
                    "deviation": 6.25,
                    "score": 1.0,
                  }
                - {
                    "average": 1,
                    "card": "Giant's Blood",
                    "deviation": 0.0,
                    "score": 1.0,
                  }
                - {
                    "average": 2,
                    "card": "The Parthenon",
                    "deviation": 0.81,
                    "score": 1.0,
                  }
                - {
                    "average": 1,
                    "card": "Archon Investigation",
                    "deviation": 0.0,
                    "score": 0.9,
                  }
                - {
                    "average": 5,
                    "card": "Villein",
                    "deviation": 1.7,
                    "score": 0.9,
                  }
                - {
                    "average": 1,
                    "card": "Wider View",
                    "deviation": 0.42,
                    "score": 0.9,
                  }
                - {
                    "average": 3,
                    "card": "Aksinya Daclau",
                    "deviation": 0.43,
                    "score": 0.8,
                  }
                - {
                    "average": 2,
                    "card": "Dreams of the Sphinx",
                    "deviation": 0.6,
                    "score": 0.8,
                  }
                - {
                    "average": 8,
                    "card": "Liquidation",
                    "deviation": 2.42,
                    "score": 0.8,
                  }
                - {
                    "average": 1,
                    "card": "Pentex™ Subversion",
                    "deviation": 0.48,
                    "score": 0.8,
                  }
        "400":
          description: |
            Malformed body, typically an invalid card name or ID
        "404":
          description: Not found, typically not enough matching examples in the TWDA
  /submit-ruling/{card_name_or_id}:
    post:
      tags: [Card]
      summary: Submit a new ruling for examination
      parameters:
        - name: card_name_or_id
          in: path
          required: true
          description: "Card name or ID"
          schema:
            type: string
            description: "Card name or ID"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  description: "Envisioned text for the ruling"
                  type: string
                link:
                  description: "Link to the official ruling from a Rules Director"
                  type: string
                  format: uri
              example:
                text: Cannot be played by vampires who aren't controlled.
                link: https://groups.google.com/d/msg/rec.games.trading-cards.jyhad/80KRDjVFkyg/TBJso9Ra_9QJ
      responses:
        "201":
          description: Successful, request created
        "400":
          description: Malformed body
        "404":
          description: Card not found
  /twda:
    post:
      tags: [Deck]
      summary: Get TWDA decks matching given filters
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                card_names:
                  description: |
                    Return decks containing those cards.
                    If absent, all decks are returned.
                  type: array
                  items:
                    description: Case insensitive card name.
                    type: string
                player:
                  description: |
                    Player name (case insensitive)
                  type: string
                date_from:
                  description: Only consider decks from this date on.
                  type: string
                  format: date
                  default: "2008-01-01"
                date_to:
                  description: Only consider decks until this date.
                  type: string
                  format: date
                  default: null
                players_count:
                  description: |
                    Only consider tournaments with at least this number of players.
                  type: integer
                  minimum: 0
                  default: 0
              example:
                cards: ["François Villon"]
                date_from: "1994-01-01"
                players: 25
      responses:
        "200":
          description: Successful, returns a list of decks as a JSON array.
          content:
            application/json:
              schema:
                type: array
                minLength: 1
                items: { $ref: "#/components/schemas/deck" }
              examples:
                { deck-array: { $ref: "#/components/examples/deck-array" } }
        "400":
          description: Malformed body
        "404":
          description: Not found
  /twda/list:
    post:
      tags: [Deck]
      summary: Get list of TWDA decks matching given filters
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                card_names:
                  description: |
                    Return a list of decks containing those cards.
                  type: array
                  items:
                    description: Case insensitive card name.
                    type: string
                player:
                  description: |
                    Player name (case insensitive)
                  type: string
                date_from:
                  description: Only consider decks from this date on.
                  type: string
                  format: date
                  default: "2008-01-01"
                date_to:
                  description: Only consider decks until this date.
                  type: string
                  format: date
                  default: null
                players_count:
                  description: |
                    Only consider tournaments with at least this number of players.
                  type: integer
                  minimum: 0
                  default: 0
              example:
                cards: ["François Villon"]
                date_from: "1994-01-01"
                players: 25
      responses:
        "200":
          description: Success, returns the list as JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck_list" }
              examples: { $ref: "#/components/examples/deck-list" }
        "400":
          description: Malformed body
        "404":
          description: Not found
  /twda/random:
    post:
      tags: [Deck]
      summary: Get a random TWDA deck matching given filters
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                card_names:
                  description: |
                    Return a random decks containing those cards.
                  type: array
                  items:
                    description: Case insensitive card name.
                    type: string
                player:
                  description: |
                    Player name (case insensitive)
                  type: string
                date_from:
                  description: Only consider decks from this date on.
                  type: string
                  format: date
                  default: "2008-01-01"
                date_to:
                  description: Only consider decks until this date.
                  type: string
                  format: date
                  default: null
                players_count:
                  description: |
                    Only consider tournaments with at least this number of players.
                  type: integer
                  minimum: 0
                  default: 0
              example:
                cards: ["François Villon"]
                date_from: "1994-01-01"
                players: 25
      responses:
        "200":
          description: Successful, returns a random deck as a JSON array.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck" }
              examples: { deck: { $ref: "#/components/examples/deck" } }
        "400":
          description: Malformed body
        "404":
          description: Not found
  /twda/{id}:
    get:
      tags: [Deck]
      summary: Get a deck by its TWDA ID
      description: Returns the deck with the given ID in the TWDA.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: "TWDA ID"
      responses:
        "200":
          description: Success, returns the deck as JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck" }
              examples: { deck: { $ref: "#/components/examples/deck" } }
        "404":
          description: Not Found
  /convert/{format}:
    post:
      tags: [Deck]
      summary: Convert a deck list format
      description: |
        Provide a text/plain or application/json deck list,
        returns the decklist in the requested format (defaults to json).
      parameters:
        - name: format
          required: true
          in: path
          schema:
            type: string
            enum: ["twd", "lackey", "json", "jol"]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/deck" }
            examples: { deck: { $ref: "#/components/examples/deck" } }
          text/plain:
            schema:
              type: string
            example: |
              Crypt (12 cards, min=7, max=24, avg=3.75)
              -----------------------------------------
              1x Gilbert Duane          7 AUS DOM OBF      prince  Malkavian:1
              1x Mariel, Lady Thunder   7 DOM OBF aus tha          Malkavian:1
              1x Badr al-Budur          5 OBF cel dom qui          Assamite:2
              1x Count Ormonde          5 OBF dom pre ser          Follower of Set:2
              1x Didi Meyers            5 DOM aus cel obf          Malkavian:1
              1x Zebulon                5 OBF aus dom pro          Malkavian:1
              1x Dimple                 2 obf                      Nosferatu:1
              1x Mustafa Rahman         2 dom                      Tremere:2
              1x Normal                 2 obf                      Malkavian:1
              1x Ohanna                 2 dom                      Malkavian:2
              1x Samson                 2 dom                      Ventrue antitribu:2
              1x Basil                  1 obf                      Pander:2

              Library (87 cards)
              Master (19; 3 trifle)
              1x Channel 10
              2x Charisma
              1x Creepshow Casino
              1x KRCG News Radio
              2x Perfectionist
              6x Storage Annex
              3x Sudden Reversal
              3x Vessel

              Ally (12)
              1x Carlton Van Wyk
              1x Gregory Winter
              1x Impundulu
              1x Muddled Vampire Hunter
              1x Ossian
              6x Procurer
              1x Young Bloods

              Equipment (9)
              1x Deer Rifle
              8x Flash Grenade

              Action Modifier (19)
              6x Cloak the Gathering
              7x Conditioning
              2x Lost in Crowds
              4x Veil the Legions

              Reaction (16)
              7x Deflection
              2x Delaying Tactics
              7x On the Qui Vive

              Combat (8)
              8x Concealed Weapon

              Event (4)
              1x FBI Special Affairs Division
              1x Hunger Moon
              1x Restricted Vitae
              1x Unmasking, The
      responses:
        "200":
          description: Success, returns the deck in given format
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck" }
              examples: { deck: { $ref: "#/components/examples/deck" } }
            text/plain:
              schema:
                type: string
              example: |
                1	Channel 10
                2	Charisma
                1	Creepshow Casino
                1	KRCG News Radio
                2	Perfectionist
                6	Storage Annex
                3	Sudden Reversal
                3	Vessel
                1	Carlton Van Wyk
                1	Gregory Winter
                1	Impundulu
                1	Muddled Vampire Hunter
                1	Ossian
                6	Procurer
                1	Young Bloods
                1	Deer Rifle
                8	Flash Grenade
                6	Cloak the Gathering
                7	Conditioning
                2	Lost in Crowds
                4	Veil the Legions
                7	Deflection
                2	Delaying Tactics
                7	On the Qui Vive
                8	Concealed Weapon
                1	FBI Special Affairs Division
                1	Hunger Moon
                1	Restricted Vitae
                1	Unmasking, The
                Crypt:
                1	Gilbert Duane
                1	Mariel, Lady Thunder
                1	Badr al-Budur
                1	Count Ormonde
                1	Didi Meyers
                1	Zebulon
                1	Dimple
                1	Mustafa Rahman
                1	Normal
                1	Ohanna
                1	Samson
                1	Bas
        "400":
          description: |
            Malformed body. Note the text parser won't fail easily:
            in case of parsing error it will list any non-parsed line as comment.
  /amaranth:
    post:
      tags: [Deck]
      summary: Retrieve a deck from Amaranth
      description: |
        Retrieve a deck from [Amaranth](https://amaranth.vtes.co.nz/) 
        using a share URL.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: "[Amaranth](https://amaranth.vtes.co.nz/) share URL"
                  format: uri
            example:
              url: "https://amaranth.vtes.co.nz/#deck/4d3aa426-70da-44b7-8cb7-92377a1a0dbd"
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: "[Amaranth](https://amaranth.vtes.co.nz/) share URL"
                  format: uri
            example:
              url: "https://amaranth.vtes.co.nz/#deck/4d3aa426-70da-44b7-8cb7-92377a1a0dbd"
      responses:
        "200":
          description: Success, returns the deck
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck" }
              examples:
                { deck: { $ref: "#/components/examples/first-blood-tremere" } }
        "400":
          description: |
            Malformed body. The Amaranth URL is expected to begin with
            https://amaranth.vtes.co.nz/#deck/
  /vdb:
    post:
      tags: [Deck]
      summary: Retrieve a deck from VDB
      description: |
        Retrieve a deck from [VDB](https://vdb.smeea.casa) 
        using a share URL.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: "[VDB](https://https://vdb.smeea.casa) share URL"
                  format: uri
            example:
              url: "https://vdb.smeea.casa/decks?id=b798e734fff7404085f7b01ad2ccb479"
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: "[VDB](https://vdb.smeea.casa) share URL"
                  format: uri
            example:
              url: "https://vdb.smeea.casa/decks?id=b798e734fff7404085f7b01ad2ccb479"
      responses:
        "200":
          description: Success, returns the deck
          content:
            application/json:
              schema: { $ref: "#/components/schemas/deck" }
              examples:
                { deck: { $ref: "#/components/examples/first-blood-tremere" } }
        "400":
          description: |
            Malformed body. The VDB URL is expected to begin with
            "https://vdb.smeea.casa/decks?id="